Extract app support modules
